# archive/idrm-mvp-docs/40-api-openapi.yaml
openapi: 3.0.3
info:
  title: IDRM API
  description: >-
    Integrated Disaster Response Management — MVP API contract.
    Canonical, incident-centric scheme under /api/v1. URIs are a stable contract
    across the SDLC: the FFP adds endpoints at these same paths, never renaming
    MVP ones. Human-readable companion: 40-api-specification.md.
  version: "1.0.0"
  contact:
    name: IDRM Development Team
  license:
    name: Proprietary (government-endorsed)
servers:
  - url: /api/v1
    description: Version-in-path base; host is environment-specific.
...
security:
  - bearerAuth: []
paths:
...
  /incidents:
    get:
      tags: [incidents]
      summary: List incidents (role-scoped map/list source, F3)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Sort' }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/IncidentStatus' }
        - in: query
          name: service_type
          schema: { $ref: '#/components/schemas/ServiceType' }
        - in: query
          name: priority
          schema: { $ref: '#/components/schemas/Priority' }
        - in: query
          name: q
          description: Free-text search over the description (PostgreSQL full-text).
          schema: { type: string, maxLength: 100 }
      responses:
        '200': { $ref: '#/components/responses/IncidentList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [incidents]
      summary: Create a help request; anonymous allowed for emergencies (F2, AC-1.4)
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentCreate' }
      responses:
        '201':
          description: Created in state `created`. `tracking_token` present for guest submissions.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Incident'
                  - type: object
                    properties:
                      tracking_token:
                        type: string
                        description: Present only for anonymous/guest submissions.
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /incidents/mine:
    get:
      tags: [incidents]
      summary: Requester's own requests (AC-7.1)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Sort' }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/IncidentStatus' }
      responses:
        '200': { $ref: '#/components/responses/IncidentList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /incidents/assigned:
    get:
      tags: [incidents]
      summary: Provider's claimed requests
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Sort' }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/IncidentStatus' }
      responses:
        '200': { $ref: '#/components/responses/IncidentList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /incidents/nearby:
    get:
      tags: [incidents]
      summary: Proximity search, nearest first (AC-4.1)
      parameters:
        - { $ref: '#/components/parameters/Latitude' }
        - { $ref: '#/components/parameters/Longitude' }
        - in: query
          name: radius_km
          required: true
          schema: { type: number, minimum: 0.1, maximum: 200 }
        - in: query
          name: service_type
          schema: { $ref: '#/components/schemas/ServiceType' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { $ref: '#/components/responses/IncidentList' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /incidents/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [incidents]
      summary: Get one request (F3, F7)
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [incidents]
      summary: Edit description/priority before it is accepted (AC-5.3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
                priority: { $ref: '#/components/schemas/Priority' }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{id}/updates:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [incidents]
      summary: Status/audit timeline for a request (F5, F7)
      responses:
        '200':
          description: Ordered updates.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/IncidentUpdate' }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{id}/approve:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Coordinator approves a critical request (AC-8.2) [created→approved]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { note: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /incidents/{id}/reject:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Coordinator rejects a request with a reason (AC-8.3) [→rejected]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /incidents/{id}/accept:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Provider claims a request; single-claim guarantee (AC-4.3, AC-4.4) [→accepted]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { eta: { type: string, format: date-time } }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '409':
          description: Already accepted by another provider, or illegal transition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /incidents/{id}/assign:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Coordinator manually assigns a provider (F8) [→accepted]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organization_id]
              properties: { organization_id: { type: string, format: uuid } }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /incidents/{id}/start:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Assigned provider starts work (AC-5.1) [accepted→in_progress]
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /incidents/{id}/complete:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Assigned provider marks complete (AC-5.2) [in_progress→completed]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notes]
              properties:
                notes: { type: string }
                photos:
                  type: array
                  items: { type: string, format: uri }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /incidents/{id}/verify:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Requester confirms completion; rating optional (AC-7.2, AC-7.3) [completed→verified]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                review: { type: string }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /incidents/{id}/cancel:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [incidents]
      summary: Withdraw a request before delivery [→cancelled]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/IncidentOne' }
        '409': { $ref: '#/components/responses/InvalidTransition' }

  # ---------------- Organizations & resources (F3, F4, F8) ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      in: query
      name: sort
      description: "Field to sort by; prefix '-' for descending, e.g. -created_at."
      schema: { type: string, example: -created_at }
    Latitude:
      in: query
      name: latitude
      required: true
      schema: { type: number, minimum: -90, maximum: 90 }
    Longitude:
      in: query
      name: longitude
      required: true
      schema: { type: number, minimum: -180, maximum: 180 }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Semantic validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InvalidTransition:
      description: The requested lifecycle transition is not allowed from the current state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    IncidentOne:
      description: A single incident.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Incident' }
    IncidentList:
      description: A paged list of incidents.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/PaginatedEnvelope'
              - type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Incident' }
  schemas:
...
    ServiceType:
      type: string
      enum: [medical, food, rescue, water, shelter, other]
    Priority:
      type: string
      enum: [low, medium, high, critical]
    IncidentStatus:
      type: string
      enum: [created, approved, accepted, in_progress, completed, verified, cancelled, rejected]
    Location:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    PaginatedEnvelope:
      type: object
      properties:
        data: { type: array, items: {} }
        pagination: { $ref: '#/components/schemas/Pagination' }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: already_accepted }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    IncidentCreate:
      type: object
      required: [service_type, priority, description, location]
      properties:
        service_type: { $ref: '#/components/schemas/ServiceType' }
        priority: { $ref: '#/components/schemas/Priority' }
        description: { type: string }
        location: { $ref: '#/components/schemas/Location' }
        photos:
          type: array
          items: { type: string, format: uri }
    Incident:
      type: object
      properties:
        id: { type: string, format: uuid }
        service_type: { $ref: '#/components/schemas/ServiceType' }
        priority: { $ref: '#/components/schemas/Priority' }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        description: { type: string }
        location: { $ref: '#/components/schemas/Location' }
        requester_id:
          type: string
          format: uuid
          nullable: true
          description: Null for anonymous/guest submissions.
        assigned_organization_id: { type: string, format: uuid, nullable: true }
        rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    IncidentUpdate:
      type: object
      properties:
        id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        from_status: { $ref: '#/components/schemas/IncidentStatus' }
        to_status: { $ref: '#/components/schemas/IncidentStatus' }
        actor_id: { type: string, format: uuid, nullable: true }
        note: { type: string }
        created_at: { type: string, format: date-time }
...